ListManagementApp/requirements.txt
torch
torchvision
google-cloud-storage
Pillow
